refactor: use interface instead of type
