+++ project-management-app/src/styled.d.ts
@@ -1,5 +1,5 @@
 import 'styled-components';
-import { Theme } from './types';
+import { Theme } from './interfaces';
 
 declare module 'styled-components' {
   export interface DefaultTheme extends Theme {}
