+++ project-management-app/src/styles/theme.ts
@@ -1,4 +1,4 @@
-import { Theme } from '../types';
+import { Theme } from '../interfaces';
 
 export const baseTheme: Theme = {
   fontFamily: {
